[plugin]Extract reusable data source loading logic(#8305)
