+++ dolphinscheduler-datasource-plugin/dolphinscheduler-datasource-hive-common/src/main/java/org/apache/dolphinscheduler/plugin/datasource/utils/CommonUtil.java
