+++ dolphinscheduler-datasource-plugin/dolphinscheduler-datasource-hive/src/main/java/org/apache/dolphinscheduler/plugin/datasource/hive/HiveDataSourceChannel.java
@@ -26,6 +26,6 @@ public class HiveDataSourceChannel implements DataSourceChannel {
 
     @Override
     public DataSourceClient createDataSourceClient(BaseConnectionParam baseConnectionParam, DbType dbType) {
-        return new HiveDataSourceClient(baseConnectionParam, dbType);
+        return new HiveCommonDataSourceClient(baseConnectionParam, dbType);
     }
 }
