+++ dolphinscheduler-datasource-plugin/pom.xml
@@ -38,5 +38,6 @@
         <module>dolphinscheduler-datasource-postgresql</module>
         <module>dolphinscheduler-datasource-api</module>
         <module>dolphinscheduler-datasource-all</module>
+        <module>dolphinscheduler-datasource-hive-common</module>
     </modules>
 </project>
